+++ service_taxi/src/main/kotlin/org/thechance/service_taxi/data/gateway/TaxiGatewayImpl.kt
@@ -1,11 +1,12 @@
-package org.thechance.service_taxi.data
+package org.thechance.service_taxi.data.gateway
 
 import org.bson.types.ObjectId
 import org.koin.core.annotation.Single
 import org.litote.kmongo.eq
-import org.thechance.service_taxi.api.models.toCollection
-import org.thechance.service_taxi.api.models.toTaxes
-import org.thechance.service_taxi.api.models.toTaxi
+import org.thechance.service_taxi.api.models.taxi.toCollection
+import org.thechance.service_taxi.api.models.taxi.toEntity
+import org.thechance.service_taxi.data.DataBaseContainer
+import org.thechance.service_taxi.data.collection.TaxiCollection
 import org.thechance.service_taxi.data.utils.paginate
 import org.thechance.service_taxi.domain.entity.Taxi
 import org.thechance.service_taxi.domain.gateway.TaxiGateway
@@ -19,12 +20,12 @@ class TaxiGatewayImpl(container: DataBaseContainer) : TaxiGateway {
     }
 
     override suspend fun getTaxiById(taxiId: String): Taxi? {
-        return collection.findOneById(ObjectId(taxiId))?.takeIf { it.isDeleted != true }?.toTaxi()
+        return collection.findOneById(ObjectId(taxiId))?.takeIf { it.isDeleted != true }?.toEntity()
     }
 
     override suspend fun getAllTaxes(page: Int, limit: Int): List<Taxi> {
         return collection.find(TaxiCollection::isDeleted eq false)
-            .paginate(page, limit).toList().toTaxes()
+            .paginate(page, limit).toList().toEntity()
     }
 
     override suspend fun deleteTaxi(taxiId: String): Boolean {
