+++ service_taxi/src/main/kotlin/org/thechance/service_taxi/domain/entity/Trip.kt
@@ -10,9 +10,3 @@ data class Trip(
     val price: Double? = null,
     val isDeleted: Boolean? = null,
 )
-
-/*
-curd
-driver/history/
-user/{id}/history/
- */
